@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { MatCardModule } from '@angular/material/card';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 
@@ -7,6 +7,7 @@ import { Block } from '../block.interface';
 import { MockOutputService } from '../mock-output.service';
 import { OutputDisplayComponent } from './output-display.component';
 import { OutputService } from '../output.service';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('OutputDisplayComponent', () => {
   let component: OutputDisplayComponent;
@@ -16,14 +17,13 @@ describe('OutputDisplayComponent', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       declarations: [OutputDisplayComponent],
-      imports: [
-        HttpClientTestingModule,
-        MatCardModule,
-        MatSnackBarModule,
-      ],
+      imports: [MatCardModule,
+        MatSnackBarModule],
       providers: [
-        { provide: OutputService, useClass: MockOutputService }
-      ],
+        { provide: OutputService, useClass: MockOutputService },
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting()
+      ]
     });
     fixture = TestBed.createComponent(OutputDisplayComponent);
     component = fixture.componentInstance;
